+++ crates/miden-standards/src/account/components/mod.rs
@@ -48,6 +48,15 @@ static MULTISIG_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
     Library::read_from_bytes(bytes).expect("Shipped Multisig library is well-formed")
 });
 
+/// Initialize the Multisig PSM library only once.
+static MULTISIG_PSM_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
+    let bytes = include_bytes!(concat!(
+        env!("OUT_DIR"),
+        "/assets/account_components/auth/multisig_psm.masl"
+    ));
+    Library::read_from_bytes(bytes).expect("Shipped Multisig PSM library is well-formed")
+});
+
 // Initialize the NoAuth library only once.
 static NO_AUTH_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
     let bytes =
@@ -123,6 +132,11 @@ pub fn multisig_library() -> Library {
     MULTISIG_LIBRARY.clone()
 }
 
+/// Returns the Multisig PSM Library.
+pub fn multisig_psm_library() -> Library {
+    MULTISIG_PSM_LIBRARY.clone()
+}
+
 /// Returns the NoAuth Library.
 pub fn no_auth_library() -> Library {
     NO_AUTH_LIBRARY.clone()
@@ -140,6 +154,7 @@ pub enum StandardAccountComponent {
     AuthSingleSig,
     AuthSingleSigAcl,
     AuthMultisig,
+    AuthMultisigPsm,
     AuthNoAuth,
 }
 
@@ -153,6 +168,7 @@ impl StandardAccountComponent {
             Self::AuthSingleSig => SINGLESIG_LIBRARY.as_ref(),
             Self::AuthSingleSigAcl => SINGLESIG_ACL_LIBRARY.as_ref(),
             Self::AuthMultisig => MULTISIG_LIBRARY.as_ref(),
+            Self::AuthMultisigPsm => MULTISIG_PSM_LIBRARY.as_ref(),
             Self::AuthNoAuth => NO_AUTH_LIBRARY.as_ref(),
         };
 
@@ -205,6 +221,9 @@ impl StandardAccountComponent {
                 Self::AuthMultisig => {
                     component_interface_vec.push(AccountComponentInterface::AuthMultisig)
                 },
+                Self::AuthMultisigPsm => {
+                    component_interface_vec.push(AccountComponentInterface::AuthMultisigPsm)
+                },
                 Self::AuthNoAuth => {
                     component_interface_vec.push(AccountComponentInterface::AuthNoAuth)
                 },
@@ -223,6 +242,7 @@ impl StandardAccountComponent {
         Self::NetworkFungibleFaucet.extract_component(procedures_set, component_interface_vec);
         Self::AuthSingleSig.extract_component(procedures_set, component_interface_vec);
         Self::AuthSingleSigAcl.extract_component(procedures_set, component_interface_vec);
+        Self::AuthMultisigPsm.extract_component(procedures_set, component_interface_vec);
         Self::AuthMultisig.extract_component(procedures_set, component_interface_vec);
         Self::AuthNoAuth.extract_component(procedures_set, component_interface_vec);
     }