+++ crates/miden-standards/src/account/interface/extension.rs
@@ -14,6 +14,7 @@ use crate::account::components::{
     basic_fungible_faucet_library,
     basic_wallet_library,
     multisig_library,
+    multisig_psm_library,
     network_fungible_faucet_library,
     no_auth_library,
     singlesig_acl_library,
@@ -111,6 +112,10 @@ impl AccountInterfaceExt for AccountInterface {
                 AccountComponentInterface::AuthMultisig => {
                     component_proc_digests
                         .extend(multisig_library().mast_forest().procedure_digests());
+                },
+                AccountComponentInterface::AuthMultisigPsm => {
+                    component_proc_digests
+                        .extend(multisig_psm_library().mast_forest().procedure_digests());
                 },
                 AccountComponentInterface::AuthNoAuth => {
                     component_proc_digests